requirements.txt
langgraph>=0.2
langchain-core>=0.3
pydantic>=2.7
python-dotenv>=1.0
openai>=1.40
google-genai>=1.0.0
httpx>=0.27
trafilatura>=1.9
selectolax>=0.3.21
playwright>=1.46
